Agent_Frameworks/Microsoft_Agent_Framework/requirements.txt
# ── Microsoft Agent Framework ─────────────────────────────────────────────────
# Requires Python >=3.10. Use: python -m venv .venv && .venv\Scripts\activate
# Then: pip install -r requirements.txt
agent-framework>=1.0.0
agent-framework-openai>=1.0.0

# ── OpenAI ────────────────────────────────────────────────────────────────────
openai>=1.0.0

# ── Search & scraping tools ───────────────────────────────────────────────────
ddgs>=0.1.0
arxiv>=2.1.0
wikipedia>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── Configuration & environment ───────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Terminal display ──────────────────────────────────────────────────────────
rich>=13.0.0

# ── Async HTTP ────────────────────────────────────────────────────────────────
aiohttp>=3.9.0
